rl-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pymongo==4.7.2
motor==3.4.0
torch==2.3.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
